chore: always tag the discovered private subnets with the cluster name (#15)
in the past, we were only doing this if the cluster name was not equal
to the install_id. we've removed the pre-emptive tagging from the
cloudformation code[1] and must therefore ensure this tag is always
applied.

Note: it seems we were NOT using the same condition for th epublic
subnets. this was not intended.

[1] https://github.com/nuonco/aws-cloudformation-templates/pull/12